gcc14 int-conversion: record ctor-result temp's concrete heap-ptr type
A parametric heap-ADT constructor (`ctor_Line`) returns the concrete
`tur_adt_Line *` even though its `(Line ..)` type c-names to the int64 carrier
(both type_c_name and emit_type_c_name collapse it). The local-var table
therefore recorded the ctor __auto_type temp as int64 (or not at all), so a later
`int64_t ln = __ps_209;` binder-init reverse straddle went undetected.

Reconstruct the concrete `tur_adt_<ctor> *` from the ctor's own name at the
call-hoist site (guarded on the ADT resolving to a heap struct/ADT) and record
that, so the binder-init bridge fires.

Fixes van-laarhoven-lens-compose (1->0) and van-laarhoven-lens-wide-compose
(1->0). Snapshot regenerated. Full suite: 2202 passed, 0 failed.

+++ tests/fixtures/van-laarhoven-lens-wide-compose/expected.c
@@ -7475,7 +7475,7 @@ int main(int argc, char **argv) {
             /* panic-return-signal: ret ctype unknown; no propagation here */
             __auto_type __ps_210 = (ctor_Line((int64_t)(intptr_t)(__ps_208), (int64_t)(intptr_t)(__ps_209)));
             /* panic-return-signal: ret ctype unknown; no propagation here */
-            int64_t ln_1364 = __ps_210;
+            int64_t ln_1364 = (int64_t)(intptr_t)(__ps_210);
             (void)ln_1364;
             __auto_type __ps_211 = (view((tur_poly_fn_t){ NULL, (int64_t(*)(void*,int64_t))__poly_1369 }, ln_1364));
             /* panic-return-signal: ret ctype unknown; no propagation here */
